jpacontroller 클래스 addrcontroller에 기능 병합
내부 로직은 동일.
단일책임 원칙으로 기능 병합
